Updated test.php to cover reminders apis

+++ api/delete_reminder.php
@@ -11,7 +11,7 @@ if (!isset($_SESSION[USER_ID]) && !TEST_MODE) {
     array_push($errors["errors"], ERROR_MISSING_LOGGED_IN_USER);
 }
 
-if (!isset($_POST['reminder_id'])) {
+if (!isset($_POST['reminder-id'])) {
 
     array_push($errors["errors"], ERROR_MISSING_VALUE . ": Reminder");
 }
@@ -20,19 +20,21 @@ if (empty($errors['errors'])) {
 
     $lookups = array();
     $user_id = !TEST_MODE ? $_SESSION[USER_ID] : 1;
-    $reminder_id = $_POST['reminder_id'];
+    $reminder_id = $_POST['reminder-id'];
 
     $user = $entityManager->find("User", $user_id);
+    $reminder = $entityManager->find("Reminder", $reminder_id);
 
-    $lookups["id"] = $id;
-    $lookups["user"] = $user;
-
-    $reminder = $entityManager->getRepository("Reminder")->findOneBy($lookups);
-
-    echo json_encode(array("success" => "Reminder deleted for " . $remind_date->format("M-d-Y h:i"), "reminder" => parse_reminder($reminder)));
+    if (!$reminder->get_user()->get_id() == $user->get_id()) {
+        array_push($errors["errors"], ERROR_INVALID_ACCESS);
+        echo json_encode($errors);
+        exit;
+    }
 
     $entityManager->remove($reminder);
+    echo json_encode(array("success" => "Reminder deleted for " . $reminder->get_remind_date()->format("M-d-Y h:i"), "reminder" => parse_reminder($reminder)));
     $entityManager->flush();
+
 
 } else {
